Fix circular dependencies + refactor messages
