+++ src/utils/ui.tsx
@@ -1,7 +1,6 @@
 import { toast } from 'react-toastify';
 import React from 'react';
 import { VideoChatData } from '../typings/interfaces';
-import { isConnected } from './general';
 
 export function setThemeColor(color: string): void {
   var themeColor: string;
@@ -39,71 +38,6 @@ export function setThemeColor(color: string): void {
   var style = document.createElement('style');
   document.head.appendChild(style);
   style.sheet?.insertRule(`:root { --themeColor: ${themeColor}}`);
-}
-
-export function displayWelcomeMessage(
-  sessionKey: string,
-  connected: Map<string, boolean>,
-  cstmSnackbarMsg?: string | HTMLElement | Element | undefined
-): void {
-  if (!isConnected(connected)) {
-    toast(
-      () => (
-        <div className="text-center justify-between">
-          {cstmSnackbarMsg ? (
-            cstmSnackbarMsg
-          ) : (
-            <>
-              <span>
-                Room ready! Waiting for others to join with session key{' '}
-              </span>
-              <strong>{sessionKey}</strong>
-            </>
-          )}
-        </div>
-      ),
-      {
-        autoClose: false,
-        toastId: 'peer_prompt',
-      }
-    );
-  }
-}
-
-export function displayWebcamErrorMessage(
-  connected: Map<string, boolean>
-): void {
-  if (!isConnected(connected)) {
-    toast(
-      () => (
-        <div className="text-center justify-between">
-          Failed to access video, please check webcam privacy settings
-          <button
-            className="snack-btn"
-            onClick={() => {
-              window.open(
-                'https://docs.catalyst.chat/docs-permissions',
-                '_blank'
-              );
-            }}
-          >
-            Help & Directions
-          </button>
-        </div>
-      ),
-      {
-        autoClose: false,
-        toastId: 'webcam/audio_error',
-      }
-    );
-  }
-}
-
-export function displayMessage(msg: string, displayLength?: number): void {
-  toast(() => <div className="text-center justify-between">{msg}</div>, {
-    toastId: 'info_msg',
-    autoClose: displayLength ? displayLength : 5000,
-  });
 }
 
 export function closeAllToasts(): void {
@@ -168,39 +102,6 @@ export function setWidth(width: number, margin: number): void {
     RemoteVideos[s].style.margin = margin + 'px';
     RemoteVideos[s].style.height = width * 0.75 + 'px';
   }
-}
-
-export function handlereceiveMessage(msg: string, color?: string): void {
-  displayMsg(msg, color ? color : 'var(--themeColor)', false);
-  document
-    .getElementById('chat-end')
-    ?.scrollIntoView({ behavior: 'smooth', block: 'nearest', inline: 'start' });
-}
-
-export function displayMsg(
-  msg: string,
-  border: string,
-  isOwnMessage: boolean
-): void {
-  if (msg.length > 0)
-    if (isOwnMessage)
-      document
-        .querySelector('.chat-messages')
-        ?.insertAdjacentHTML(
-          'beforeend',
-          `<div class="message-item customer cssanimation fadeInBottom"><div class="message-bloc" style="border: 3px solid ${border}"><div class="message">` +
-            msg +
-            '</div></div></div>'
-        );
-    else
-      document
-        .querySelector('.chat-messages')
-        ?.insertAdjacentHTML(
-          'beforeend',
-          `<div class="message-item moderator cssanimation fadeInBottom"><div class="message-bloc" style="border: 3px solid ${border}"><div class="message">` +
-            msg +
-            '</div></div></div>'
-        );
 }
 
 export function uuidToHue(uuid: string, VC: VideoChatData): number {
